RFPs Analyzer/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==4.0.0
groq-sdk==0.1.0
openai==1.12.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
pinecone==0.2.0
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.1.0
pytesseract==0.3.10
Pillow==10.0.0
python-magic==0.4.27
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.4
pinecone-client==3.0.2
pydantic==2.5.2